requirements.txt
# Python 3.11+ — yonnn bug bounty framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy[asyncio]>=2.0.36,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.14.0,<2.0.0
celery>=5.4.0,<6.0.0
redis>=5.2.0,<6.0.0
pydantic-settings>=2.6.0,<3.0.0
loguru>=0.7.2,<1.0.0
httpx>=0.28.0,<1.0.0
bcrypt>=4.2.0,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.9,<1.0.0
email-validator>=2.2.0,<3.0.0
